add html writer, replace it in readme_preview and readme
